requirements.txt
streamlit>=1.34.0
pydantic>=2.6.0
pandas>=2.0.0
langgraph>=0.2.0
openai>=1.30.0
python-dotenv>=1.0.0
